# app/swagger.yaml
swagger: "2.0"
info:
  title: "Banking API"
  description: "API documentation for the Banking application."
  version: "1.0.0"
host: "localhost:5000"
basePath: "/api"
schemes:
  - "http"
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Enter your JWT token in the format **Bearer &lt;token&gt;**"
paths:
...
  /accounts/{account_id}/transactions:
    get:
      summary: "Get Account Transactions"
      description: "Retrieve a list of transactions for a specific account."
      tags:
        - "Accounts"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "account_id"
          required: true
          type: "integer"
          description: "ID of the account to retrieve transactions for"
        - in: "query"
          name: "start_date"
          type: "string"
          format: "date"
          description: "Filter transactions from this date (YYYY-MM-DD)"
          example: "2023-01-01"
        - in: "query"
          name: "end_date"
          type: "string"
          format: "date"
          description: "Filter transactions up to this date (YYYY-MM-DD)"
          example: "2023-12-31"
        - in: "query"
          name: "type"
          type: "string"
          description: "Filter transactions by type (deposit, withdrawal, transfer)"
          example: "deposit"
        - in: "query"
          name: "search"
          type: "string"
          description: "Search transactions by description"
          example: "groceries"
        - in: "query"
          name: "page"
          type: "integer"
          description: "Page number for pagination"
          example: 1
        - in: "query"
          name: "per_page"
          type: "integer"
          description: "Number of transactions per page"
          example: 20
      responses:
        200:
          description: "List of transactions"
          schema:
            type: "object"
            properties:
              transactions:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "integer"
                      example: 1
                    description:
                      type: "string"
                      example: "Grocery shopping"
                    amount:
                      type: "number"
                      format: "float"
                      example: 50.0
                    timestamp:
                      type: "string"
                      format: "date-time"
                      example: "2023-01-01T12:00:00Z"
              pg:
                type: "integer"
                example: 1
              per_pg:
                type: "integer"
                example: 20
              total_items:
                type: "integer"
                example: 100
        404:
          description: "Account not found"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Account not found"
  /transactions:
    get:
      summary: "Get Transaction History"
      description: "Retrieve the transaction history for all accounts belonging to the authenticated user."
      tags:
        - "Transactions"
      security:
        - Bearer: []
      responses:
        200:
          description: "List of transactions"
          schema:
            type: "object"
            properties:
              transactions:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "integer"
                      example: 1
                    transaction_type:
                      type: "string"
                      example: "deposit"
                    amount:
                      type: "number"
                      format: "float"
                      example: 100.0
                    timestamp:
                      type: "string"
                      format: "date-time"
                      example: "2023-01-01T12:00:00Z"
                    description:
                      type: "string"
                      example: "Grocery shopping"
                    from_account_id:
                      type: "integer"
                      example: 1
                    to_account_id:
                      type: "integer"
                      example: 2
  /transactions/deposit:
    post:
      summary: "Deposit Funds"
      description: "Deposit funds into a specific account."
      tags:
        - "Transactions"
      security:
        - Bearer: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "Details of the deposit transaction"
          schema:
            type: "object"
            required:
              - account_id
              - amount
            properties:
              account_id:
                type: "integer"
                example: 1
              amount:
                type: "number"
                format: "float"
                example: 100.0
              description:
                type: "string"
                example: "Deposit for savings"
      responses:
        200:
          description: "Deposit successful"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Deposit successful"
              transaction:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  transaction_type:
                    type: "string"
                    example: "deposit"
                  amount:
                    type: "number"
                    format: "float"
                    example: 100.0
                  timestamp:
                    type: "string"
                    format: "date-time"
                    example: "2023-01-01T12:00:00Z"
              new_balance:
                type: "number"
                format: "float"
                example: 500.0
  /transactions/withdraw:
    post:
      summary: "Withdraw Funds"
      description: "Withdraw funds from a specific account."
      tags:
        - "Transactions"
      security:
        - Bearer: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "Details of the withdrawal transaction"
          schema:
            type: "object"
            required:
              - account_id
              - amount
            properties:
              account_id:
                type: "integer"
                example: 1
              amount:
                type: "number"
                format: "float"
                example: 50.0
              description:
                type: "string"
                example: "Withdrawal for groceries"
      responses:
        200:
          description: "Withdrawal successful"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Withdrawal successful"
              transaction:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  transaction_type:
                    type: "string"
                    example: "withdrawal"
                  amount:
                    type: "number"
                    format: "float"
                    example: 50.0
                  timestamp:
                    type: "string"
                    format: "date-time"
                    example: "2023-01-01T12:00:00Z"
              new_balance:
                type: "number"
                format: "float"
                example: 450.0
  /transactions/transfer:
    post:
      summary: "Transfer Funds"
      description: "Transfer funds between two accounts."
      tags:
        - "Transactions"
      security:
        - Bearer: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "Details of the transfer transaction"
          schema:
            type: "object"
            required:
              - from_account_id
              - to_account_id
              - amount
            properties:
              from_account_id:
                type: "integer"
                example: 1
              to_account_id:
                type: "integer"
                example: 2
              amount:
                type: "number"
                format: "float"
                example: 200.0
              description:
                type: "string"
                example: "Transfer to savings account"
      responses:
        200:
          description: "Transfer successful"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Transfer successful"
              transaction:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  transaction_type:
                    type: "string"
                    example: "transfer"
                  amount:
                    type: "number"
                    format: "float"
                    example: 200.0
                  timestamp:
                    type: "string"
                    format: "date-time"
                    example: "2023-01-01T12:00:00Z"
              from_account_balance:
                type: "number"
                format: "float"
                example: 800.0
              to_account_balance:
                type: "number"
                format: "float"
                example: 1200.0
  /transactions/transfer-advanced:
    post:
      summary: "Advanced Transfer Funds"
      description: "Transfer funds between accounts with additional flexibility."
      tags:
        - "Transactions"
      security:
        - Bearer: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "Details of the advanced transfer transaction"
          schema:
            type: "object"
            required:
              - from_account_id
              - to_account_id
              - amount
            properties:
              from_account_id:
                type: "integer"
                example: 1
              to_account_id:
                type: "integer"
                example: 2
              amount:
                type: "number"
                format: "float"
                example: 300.0
              description:
                type: "string"
                example: "Advanced transfer to another account"
      responses:
        200:
          description: "Advanced transfer successful"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Transfer successful"
              transaction:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  transaction_type:
                    type: "string"
                    example: "transfer"
                  amount:
                    type: "number"
                    format: "float"
                    example: 300.0
                  timestamp:
                    type: "string"
                    format: "date-time"
                    example: "2023-01-01T12:00:00Z"
              from_account_balance:
                type: "number"
                format: "float"
                example: 700.0
              to_account_balance:
                type: "number"
                format: "float"
                example: 1300.0
        400:
          description: "Invalid request or insufficient funds"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Insufficient funds or invalid request"
        404:
          description: "Account not found"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Source or destination account not found"
  /transactions/accounts/{account_id}/transactions:
    get:
      summary: "Get Transactions for a Specific Account"
      description: "Retrieve a list of transactions for a specific account."
      tags:
        - "Transactions"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "account_id"
          required: true
          type: "integer"
          description: "ID of the account to retrieve transactions for"
      responses:
        200:
          description: "List of transactions for the account"
          schema:
            type: "object"
            properties:
              transactions:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "integer"
                      example: 1
                    transaction_type:
                      type: "string"
                      example: "deposit"
                    amount:
                      type: "number"
                      format: "float"
                      example: 100.0
                    timestamp:
                      type: "string"
                      format: "date-time"
                      example: "2023-01-01T12:00:00Z"
                    description:
                      type: "string"
                      example: "Deposit to savings account"
        404:
          description: "Account not found"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Account not found or does not belong to you"

    post:
      summary: "Create a Transaction for a Specific Account"
      description: "Create a new transaction (deposit, withdrawal, or transfer) for a specific account."
      tags:
        - "Transactions"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "account_id"
          required: true
          type: "integer"
          description: "ID of the account to create a transaction for"
        - in: "body"
          name: "body"
          required: true
          description: "Details of the transaction"
          schema:
            type: "object"
            required:
              - type
              - amount
            properties:
              type:
                type: "string"
                example: "deposit"
                description: "Type of transaction (deposit, withdrawal, transfer)"
              amount:
                type: "number"
                format: "float"
                example: 200.0
              to_account_id:
                type: "integer"
                example: 2
                description: "Required for transfers"
              description:
                type: "string"
                example: "Deposit to savings account"
      responses:
        201:
          description: "Transaction created successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Deposit successful"
              transaction:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  transaction_type:
                    type: "string"
                    example: "deposit"
                  amount:
                    type: "number"
                    format: "float"
                    example: 200.0
                  timestamp:
                    type: "string"
                    format: "date-time"
                    example: "2023-01-01T12:00:00Z"
              new_balance:
                type: "number"
                format: "float"
                example: 1200.0
        400:
          description: "Invalid request or insufficient funds"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Invalid transaction type or insufficient funds"
        404:
          description: "Account not found"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Account not found or does not belong to you"